FIXED Visual bugs and performance
