+++ .agents/rules/ai-engineering.md
@@ -13,3 +13,6 @@ Você está implementando integrações de Inteligência Artificial ou desenvolv
 5. **Seleção de Modelos por Tarefa**: Modele sua arquitetura com base em custo-benefício. Tarefas iterativas rápidas e simples (como sugerir nome para um botão) devem usar modelos velozes (ex: `gemini-3.5-flash`, `claude-haiku-4-5-20251001`, `gpt-5.4-mini`). Análises complexas de Root Cause Analysis ou criação de cenários Gherkin merecem modelos densos (ex: `claude-opus-4-8`, `gpt-5.4`, `gemini-3-flash-preview`). Sempre que for tomar alguma decisão de qual modelo usar para um provider, você DEVE antes rodar um comando via terminal ou script para puxar via API a lista de modelos atualizados disponíveis daquele provider, para evitar o uso de modelos depreciados.
 6. **Evite "Conversas" Desnecessárias**: O Robot Runner é uma ferramenta técnica. Os agentes autônomos criados na aplicação não devem ser "tagarelas" pedindo "por favor" ou "olá". Mantenha o formato orientado a dados, focado puramente em extração de metadados de UI, ADB e Scripts de Teste.
 7. **Evite uso de IA onde o código é a melhor opção**: O Robot Runner é uma ferramenta para uso profissional, onde ambos tempo e dinheiro são valiosos. Use a IA apenas onde for necessário, usando abordagens híbridas (código apoiado por IA se necessário) onde for possível, agilizando execuções e diminuindo gastos com tokens.
+8. **UX de Prompts Sistêmicos**: Ao disparar prompts automatizados gerados pelo Robot Runner que visam instruir agentes a realizarem tarefas com base em UI (ex: converter ações gravadas em testes), sempre marque a mensagem como `hidden` na interface do usuário (ex: no `AiAgentPanel.tsx`). O usuário não deve ver blocos enormes de texto e JSON gerados pela própria IDE; mostre apenas a resposta inteligente do modelo.
+9. **Contexto de Idioma Dinâmico**: Ao invés de usar *hardcoding* para estipular o idioma de resposta da IA dentro dos prompts gerados, sempre injete o estado de idioma atual (`settings.language`) no corpo do prompt. Isso assegura que a resposta do agente acompanhe o *locale* atual configurado pelo usuário (PT-BR, EN, ES, etc).
+10. **Limitações do PowerShell (Windows)**: Ao construir invocações CLI pelo Rust ou mesmo ao executar comandos no terminal durante o desenvolvimento, tenha cuidado com *aliases* do PowerShell. Por exemplo, `dir /b` ou `ls` com flags nativas de bash/cmd costumam falhar no PS (`dir` é alias para `Get-ChildItem`). Use ferramentas nativas (como a API fs do Tauri) ou envolva explicitamente em um invólucro do `cmd /c` se for usar flags de linha de comando legadas do DOS.
