+++ src/components/molecules/FileSavedFeedback.tsx
@@ -26,8 +26,10 @@ export function FileSavedFeedback({ path, onClose, className = "" }: FileSavedFe
             </span>
             {onClose && (
                 <Button
+                    variant="ghost"
+                    size="sm"
                     onClick={onClose}
-                    className="ml-2 hover:bg-success/10 rounded p-1 transition-colors"
+                    className="ml-2 rounded-full p-1.5 hover:bg-success/20 hover:text-success transition-colors h-auto w-auto"
                 >
                     <X size={14} />
                 </Button>
